containers/subscriber/requirements.txt
requests
google-api-python-client
configparser
google-cloud-pubsub
python-dateutil
google-cloud-language
flexmock
pytest_bdd
pytest-env
pytest-cov
testinfra
